feature/navBar: Componentização da navBar
